Don't send files twice and remove extra \r\n from header
After the manager http auth changes, we forgot to remove the manual
sending of the file. Also, ast_http_send adds two \r\n to the header that
is passed to it, so a trailing \r\n is removed from the Content-type
header. It might be better to change ast_http_send, but I don't like changing
the behavior of an API function.

(closes issue #17239)
Reported by: cjacobsen
Patches: 
      patch2.diff uploaded by cjacobsen (license 1029)
Tested by: lathama, cjacobsen



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@270660 f38db490-d61c-443f-a65b-d21fe96a405b