Make sure reload updates SLA config
Even if there are no stations or trunks defined, we need to start the sla
thread to make sure we get the reload event. Also, when doing a reload we need
to remove the existing trunks and stations or they end up hanging around.

(closes issue #16818)
Reported by: mbonin
Patches: 
      sla_reload.patch uploaded by twilson (license 396)
Tested by: twilson


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@272109 f38db490-d61c-443f-a65b-d21fe96a405b