Add option to not do a call forward on 482 Loop Detected
Asterisk has always set up a forwarded call when receiving a 482 Loop Detected.
This prevents handling the call failure by just continuing on in the dialplan.
Since this would be a change in behavior, the new option to disable this
behavior is forwardloopdetected which defaults to 'yes'.

Review: https://reviewboard.asterisk.org/r/764/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@274280 f38db490-d61c-443f-a65b-d21fe96a405b