Fix an issue where the port for p->ourip was being set to 0.
This should fix all the CDR tests that were not passing. When they would
originate a call, all fields in the INVITE that contained the source port would
have the port set to 0. Most troubling of these was the Contact header. Tests
are passing locally now and should also pass on the bamboo build agents.



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@275294 f38db490-d61c-443f-a65b-d21fe96a405b