Change ast_write to not stop generator when called from ast_prod.
For SIP channels configured with the progressinband option on, the ringback was
being immediately stopped. This problem was due to ast_prod being moved for a
deadlock fix in 259858. Prodding the channel after setting up the generator
triggered the check in ast_write to stop the generator. The fix here should
write the frame the same as was done before the call to ast_prod was moved.

(closes issue #17372)
Reported by: tech_admin


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@275665 f38db490-d61c-443f-a65b-d21fe96a405b