priexclusive in chan_dahdi.conf ignored when reloading dahdi module
During a reload, the priexclusive and outsignalling parameters are not
read in from the config file as intended.  Unfortunately, they get set to
defaults as a result.  This patch makes sure that they do not get set to
defaults during a reload.

(closes issue #17441)
Reported by: mtryfoss
Patches:
      issue17441_v1.4.patch uploaded by rmudgett (license 664)
      issue17441_v1.6.2.patch uploaded by rmudgett (license 664)
      issue17441_trunk.patch uploaded by rmudgett (license 664)
Tested by: rmudgett


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@277419 f38db490-d61c-443f-a65b-d21fe96a405b