Fix parsing error in sip_sipredirect().
The code was written in a way that did a bad job of
parsing the port out of a URI. Specifically, it would
do badly when dealing with an IPv6 address. In this
particular scenario, there was no value from parsing
the port out, so I just removed that logic. And while
I was messing around in the function, I changed some
variable names to be more descriptive.

(closes issue #17661)
Reported by: oej
Patches: 
      17661.diff uploaded by mmichelson (license 60)



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@279887 f38db490-d61c-443f-a65b-d21fe96a405b