Prevent loss of Caller ID information set on local channel after masquerade.
Caller ID set on the channel before a masquerade occurs when using a local
channel would cause the information to be lost. The problem was that the
information was set on a channel destined to be hung up. The somewhat confusing
fix is to detect if any Caller ID has been set on the channel and if so 
preswap the Caller ID data so that basically the masquerade puts the data back.

(closes issue #17138)
Reported by: kobaz

Review: https://reviewboard.asterisk.org/r/847/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@281390 f38db490-d61c-443f-a65b-d21fe96a405b