tos_sip option was not being set correctly
When tos_sip is used, the tos of the sip socket is only set
correctly if the socket binding changes on a reload.  If the binding
stays the same but the TOS changes, the new tos value would not take
into effect.  This patch fixes that.


(closes issue #17712)
Reported by: nickb


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@282893 f38db490-d61c-443f-a65b-d21fe96a405b