authenticate OPTIONS requests just like we would an INVITE
OPTIONS requests should be treated the same as an INVITE
This includes authentication.  This patch adds the ability for
incoming out of dialog OPTION requests to be authenticated
before providing a response indicating whether an extension
is available or not.  The authentication routine works the
exact same way as it does for incoming INVITEs.  This means
that if a peer has 'insecure=invite' in their peer definition,
the same will be true for the processing of the OPTIONS request.

Review: https://reviewboard.asterisk.org/r/881/



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@284950 f38db490-d61c-443f-a65b-d21fe96a405b