Rate limit calls to fsync() to 1 per second after astdb updates.
Astdb was determined to be one of the most significant bottlenecks in SIP
registration processing.  This patch improved the speed of an astdb load
test by 50000% (yes, Fifty-Thousand Percent).  On this particular load test
setup, this doubled the number of SIP registrations the server could handle.

Review: https://reviewboard.asterisk.org/r/825/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@286112 f38db490-d61c-443f-a65b-d21fe96a405b