Fix a 100% CPU consumption problem when setting console=yes in asterisk.conf.
The handling of -c and console=yes should be the same, but they were not.
When you specify -c, it sets both a flag for console module and for asterisk
not to fork() off into the background.  The handling of console=yes only set
console mode, so you would end up with a background process() trying to run
the Asterisk console and freaking out since it didn't have anything to read
input from.

Thanks to beagles for reporting and helping debug the problem!


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@288339 f38db490-d61c-443f-a65b-d21fe96a405b