Safer xml parsing, treat all clients the same, and better local candidate selection.
The gtalk channel driver was doing several unsafe operations
in regards to how it parsed incoming XML messages.  I have cleaned
that code up so it should be much safer now.

We now treat all clients types the same.  We have no reason to
distinguish between GMAIL and GOOGLE VOICE clients anymore because
they all work the same way.

I also modified how the local ip is found.  If no bindaddress is provided
in the config file, we attempt to determine the local ip we
would use to connect to google.com.  If that fails, then
we fall back to the ast_find_ourip() function as a last resort.
Using the new method makes it much less likely that we would ever
advertise a local RTP candidate as a loopback address.



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@291827 f38db490-d61c-443f-a65b-d21fe96a405b