Fix playback failure when using IAX with the timerfd module.
To fix this issue the alert pipe will now be used when the timerfd module is
in use. There appeared to be a race that was not solved by adding locking in the
timerfd module, but needed to be there anyway. The race was between the timer
being put in non-continuous mode in ast_read on the channel thread and the IAX 
frame scheduler queuing a frame which would enable continuous mode before the
non-continuous mode event was read. This race for now is simply avoided.

(closes issue #18110)
Reported by: tpanton
Tested by: tpanton

I put tested by tpanton because it was tested on his hardware. Thanks for the
remote access to debug this issue!


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.6.2@294277 f38db490-d61c-443f-a65b-d21fe96a405b