Add ability to disable T.38 usage for specific SendFAX/ReceiveFAX sessions.
Sometimes during troubleshooting it can be useful to disable T.38 usage in order
to narrow down a problem. This patch adds an 'n' option to SendFAX and ReceiveFAX
so that can be done without having to disable T.38 usage entirely for the peer
that Asterisk is communicating with.

(inspired by trying to assist Bryant Zimmerman on asterisk-users)



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@304342 f38db490-d61c-443f-a65b-d21fe96a405b