When we pass the S() or L() options to MeetMe, make sure that we honor C as well.
Without this patch, if the user was kicked from the conference via the S() or L()
mechanism, we would just hang up on them even if we also passed C (continue in
dialplan when kicked).  With this patch we honor the C flag in those cases.

(closes issue #17317)
Reported by: var


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.6.2@304773 f38db490-d61c-443f-a65b-d21fe96a405b