Don't allow a REFER w/replaces to replace its own dialog
Asterisk currently accepts a REFER with a Refer-To with an embedded Replaces
header that matches the dialog of the REFER. This would be a situation like A
calls B, A calls C, A transfers B to A, which is just silly. This patch makes
the transfer fail instead of making Asterisk freak out and forget to hang other
channels up.

Review: https://reviewboard.asterisk.org/r/1093/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@306617 f38db490-d61c-443f-a65b-d21fe96a405b