Don't try to pickup a call in the middle of a masquerade
If A calls B which doesn't answer and C & D both try to do a call pickup, it is
possible for ast_pickup_call to answer the call, then fail to masquerade one of
the calls because the other one is already in the process of masquerading. This
patch checks to see if the channel is in the process of masquerading before
call before selecting it for a pickup.

Review: https://reviewboard.asterisk.org/r/1094/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@306672 f38db490-d61c-443f-a65b-d21fe96a405b