Calling a gosub routine defined in AEL from Dial/Queue ceased to work.
A bug in AEL did not distinguish between the "s" extension generated by
AEL and an "s" extension that was required to exist by the chan_dahdi
(or another channel) that was not supplied with a starting extension.
Therefore, AEL made incorrect assumptions about what commands were
permissable in the context.  This was fixed by making AEL generate a
different extension name.  However, Dial and Queue make additional
assumptions about the name of the default gosub extension.  Therefore,
they needed to be brought into line with a "macro" rendered by AEL (as
a gosub), without breaking traditional dialplans written without the
aid of AEL.

Related to (issue #18480)
 Reported by: nivek

(closes issue #18729)
 Reported by: kkm
 Patches: 
       20110209__issue18729.diff.txt uploaded by tilghman (license 14)
       018729-dial-queue-gosub-try3.patch uploaded by kkm (license 888)
 Tested by: kkm


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@307750 f38db490-d61c-443f-a65b-d21fe96a405b