Add setvar option to calendaring
Adding the setvar option with variable substitution on the value allows things
like setting the outbound caller id name to the summary of a calendar event,
etc. Values could be chained together as they are appended in order to do some
scripting if necessary.

Review: https://reviewboard.asterisk.org/r/1134/


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@309640 f38db490-d61c-443f-a65b-d21fe96a405b