Add setsubstate_busy.
Move handling of setting busy state from skinny_indicate to it's own sub.
Also, modified behaviour to not hangup the sub and let the dialplan
have a chance in doing what it wants (eg busy(10); hangup() in the dialplan
now gives a busy indication for 10 secs and then hangs up.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@316994 f38db490-d61c-443f-a65b-d21fe96a405b