Fix up skinny hints.
Probably haven't been working for a couple of years. May still need
some more love, but they are now working, both as a hint device and
monitoring a hint. Changes centre around the long ago change
to remove the requirement for a device name in a skinny line, and 
changes to the transmit_* functions.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@319316 f38db490-d61c-443f-a65b-d21fe96a405b