Fixes an imapfolder related crash
imapfolders being set in the general section of voicemail would cause the inbox folder name to
change.  Since sound file names are made based on the names of the folders, this would cause
the audio related to that folder name to change and if Asterisk attempted to play it, the
channel would instantly hang up when the audio file couldn't be found.  This patch searches for
the name of the folder first to leave existing behavior in tact and if that fails, it uses
the normal inbox name to get the sound file instead.


(closes issue #16104)
Reported by: blkline

Review: https://reviewboard.asterisk.org/r/1215/


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@320162 f38db490-d61c-443f-a65b-d21fe96a405b