Add ConnectedLineNum/Name headers to output of AMI action Status.
* Add ConnectedLineNum and ConnectedLineName headers to the output of the
AMI action Status.  This makes it easier to find out who the channel is
connected to without having to lookup BridgedChannel or when they are
connected to an application (e.g.: VoiceMail) which has no bridged
channel.

* Bridged channels with no CallerID had "" instead of "<unknown>" output,
that might be a bug as "<unknown>" was what older versions used.

(closes issue #18158)
Reported by: gareth
Patches:
      svn-292308.diff uploaded by gareth (license 208)


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.8@320650 f38db490-d61c-443f-a65b-d21fe96a405b